Added mongoDB setup and parsing changes
mongoDB is launched with a subprocess and can be killed using a process
group id.  OSM json data imported using mongoimport, which is also
called with a subprocess.

Made changes to shape_element to clean code and account for nodes and
ways that have tag children with “k” = “type” (these were overwriting
the “type” key in the returned element).